requirements.txt
coverage==4.4.1
Django==1.11
django-cors-headers==2.1.0
django-nose==1.4.4
djangorestframework==3.6.3
gitdb2==2.0.2
GitPython==2.1.5
nose==1.3.7
pytz==2017.2
smmap2==2.0.3
